+++ src/main/java/hexlet/code/service/UserService.java
@@ -3,9 +3,11 @@ package hexlet.code.service;
 import hexlet.code.dto.user.UserCreateDTO;
 import hexlet.code.dto.user.UserDTO;
 import hexlet.code.dto.user.UserUpdateDTO;
+import hexlet.code.exception.ForbiddenOperationException;
 import hexlet.code.exception.ResourceNotFoundException;
 import hexlet.code.mapper.UserMapper;
 import hexlet.code.model.User;
+import hexlet.code.repository.TaskRepository;
 import hexlet.code.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
@@ -21,6 +23,9 @@ public class UserService implements UserDetailsService {
 
     @Autowired
     private UserRepository userRepository;
+
+    @Autowired
+    private TaskRepository taskRepository;
 
     @Autowired
     private UserMapper userMapper;
@@ -59,6 +64,9 @@ public class UserService implements UserDetailsService {
     }
 
     public void delete(Long id) {
+        if (!taskRepository.findAllByAssigneeId(id).isEmpty()) {
+            throw new ForbiddenOperationException("Forbidden operation");
+        }
         userRepository.deleteById(id);
     }
